Put parking_lot_count into the same folder as parking_lot_detect
